save currency values as decimals

+++ server/migrations/20191106112856_transactions.up.sql
@@ -1,9 +1,9 @@
 CREATE TABLE transactions
 (
     `id`         integer PRIMARY KEY NOT NULL AUTO_INCREMENT,
-    `new_saldo`  float               NOT NULL,
-    `old_saldo`  float               NOT NULL,
-    `amount`     float               NOT NULL,
+    `new_saldo`  decimal(15, 2)      NOT NULL,
+    `old_saldo`  decimal(15, 2)      NOT NULL,
+    `amount`     decimal(15, 2)      NOT NULL,
     `account_id` integer             NOT NULL,
     `created`    datetime            NOT NULL
 );
@@ -12,5 +12,5 @@ ALTER TABLE `transactions`
     ADD FOREIGN KEY (`account_id`) REFERENCES `accounts` (`id`);
 
 
-CREATE INDEX idx_created ON transactions(`created`);
-CREATE INDEX idx_id ON transactions(`id`)
+CREATE INDEX idx_created ON transactions (`created`);
+CREATE INDEX idx_id ON transactions (`id`)
